refactor(runners): extract the residency daemon out of claude_session (v4 step 3)
Pure extraction, zero behavior change. The daemon machinery — pid file,
signal handlers, heartbeat side-task, turn inbox, a2a sidecar spawn,
autonomous drive-until-done loop, and the shutdown sequence — moves
verbatim from _runners/claude_session.py into _runners/session_daemon.py
as run_session_daemon(name, *, turn_driver, ...). Per the v4 design,
sac owns the process; a harness owns only the turn: the daemon is
vendor-blind and receives the conversation as a required turn_driver
callable.

claude_session.run keeps its exact signature and becomes a thin wrapper
that resolves the Claude turn driver (run_conversation_fn seam, default
_session_conversation.run_conversation) and delegates. _autonomous_loop
and _heartbeat_loop stay importable from claude_session so every
existing runner.<name> call shape survives.
